Fix imported assets animations not appearing bug
Rest of the fix is in the editor side

Task-number: QT3DS-3921

+++ src/dm/systems/Qt3DSDMAnimation.h
@@ -41,7 +41,7 @@ struct SLinearKeyframe
     float m_value;
 
     SLinearKeyframe() = default;
-    SLinearKeyframe(long time, float value)
+    SLinearKeyframe(float time, float value)
     : m_time(time)
     , m_value(value) {}
 };
